requirements.txt
dotenv
requests
flask
beautifulsoup4
pathlib
streamlit
pandas
openai
matplotlib
numpy
pyautogen
yfinance

# huggingface_hub 
# transformers 
# diffusers 
# datasets 
# evaluate 
# accelerate
# ipywidgets

